requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
cloudinary==1.36.0
Django==4.2.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
environs==9.5.0
gunicorn==21.2.0
idna==3.4
marshmallow==3.20.1
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.6
whitenoise==6.5.0
